+++ CHANGES.md
@@ -13,6 +13,7 @@
 - Implemented optimized qubit kernels for diagonal gates and controlled
   diagonal gates: `apply_diag()` and `applyCTRL_diag()`, along with their
   in-place counterparts, `apply_diag_inplace()` and `applyCTRL_diag_inplace()`
+- Added [`examples/qft_inplace.cpp`](examples/qft_inplace.cpp) example
 
 ## Version 7.0.3 - 25 March 2026